+++ cartridge/webui/api-topology.lua
@@ -273,10 +273,6 @@ local function set_failover_enabled(_, args)
     return admin.set_failover_enabled(args.enabled)
 end
 
-local function set_demo_uri(_, args)
-    return admin.set_demo_uri(args.demo_uri)
-end
-
 local function init(graphql)
 
     graphql.add_callback({
@@ -406,7 +402,7 @@ local function init(graphql)
     graphql.add_callback({
         prefix = 'cluster',
         name = 'self',
-        doc = 'Get current server',
+        doc = 'Some information about current server',
         args = {},
         kind = gql_types.object({
             name = 'ServerShortInfo',
@@ -440,17 +436,6 @@ local function init(graphql)
         }),
         callback = module_name .. '.edit_topology',
     })
-
-    graphql.add_mutation({
-        prefix = 'cluster',
-        name = 'self',
-        doc = 'Set demo uri variable',
-        args = {
-            demo_uri = gql_types.string,
-        },
-        kind = gql_types.string,
-        callback = module_name .. '.set_demo_uri',
-    })
 end
 
 return {
@@ -474,6 +459,4 @@ return {
     get_known_roles = get_known_roles,
     get_failover_enabled = get_failover_enabled,
     set_failover_enabled = set_failover_enabled,
-
-    set_demo_uri = set_demo_uri,
 }